drivers/ldr, drivers/tc1047: updates after test
git-svn-id: http://svn.arune.se/svn/HomeAutomation/trunk@799 e69ec2e4-db17-0410-bd4a-d6f28f89cd0d

+++ EmbeddedSoftware/AVR/avr-lib/drivers/sensor/ldr/ldr.c
@@ -30,7 +30,7 @@ uint8_t LDR_SensorInit(void)
 /*
  * Start reading sensor value
  */
-uint32_t LDR_GetData(uint8_t sensor)
+uint16_t LDR_GetData(uint8_t sensor)
 {
     return ADC_Get(sensor);
 }
